requirements.txt
numpy
pandas
opencv-contrib-python
streamlit
pillow
scipy
joblib
scikit-image